Moved Markdown.jsx to templates directory
